Move tracker logic to separate file (#478)
* Starting namespace add

* Adding namespace to root class

* Working on tests

* Working on UI tests

* Address StructuredData tests

* Explicit WP_User import

* Fix linting issues

* Using fully qualified names on docstrings in StructuredData tests

* Correcting more docstrings

* Fixing warnings on UI tests

* Fix get_options

* Removing warnings on AdminWarningTest

* Fix FBIA tests

* Fix Integrations Test

* Splitting out tracker code

* Adding tracker test

* Adding additional tests

* Keep moving tests

* Moving constant

* Adding WP_Scripts

* Renaming class to scripts

* Fixing scripts naming

* Fix path

* Making sure script enqueue is clean

* Fix linting issues

* Updating comment

* Updating more comments
